+++ src/settings/base.py
@@ -413,14 +413,21 @@ SUBSCRIPTION_PLANS = {
     "BASIC": {
         "fare": 0,  # Cents.
         "attachments": {
+            "documents": {
+                "max_file_size":        5242800,
+                "max_per_event":        5,
+                "max_per_organization": 5,
+            },
             "images": {
                 "max_width":            1600,
                 "max_height":           900,
+                "max_file_size":        10485760,
                 "max_per_event":        25,
                 "max_per_organization": 25,
                 "quality":              80,
             },
-            "documents": {
+            "video": {
+                "max_file_size":        10485760,
                 "max_per_event":        5,
                 "max_per_organization": 5,
             },
@@ -435,32 +442,39 @@ SUBSCRIPTION_PLANS = {
         },
         "accounts": {},
         "events": {
-            "upon_request_only":    False,
             "max_per_day":          1,
             "max_per_week":         None,
             "max_per_month":        None,
             "max_per_year":         None,
+            "upon_request_only":    False,
         },
         "organizations": {
-            "upon_request_only":    True,
             "max_per_day":          0,
             "max_per_week":         None,
             "max_per_month":        None,
             "max_per_year":         None,
+            "upon_request_only":    True,
         },
         "places": {},
     },
     "TIER-1": {
         "fare": 0,  # Cents.
         "attachments": {
+            "documents": {
+                "max_file_size":        5242800,
+                "max_per_event":        5,
+                "max_per_organization": 5,
+            },
             "images": {
                 "max_width":            1920,
                 "max_height":           1080,
+                "max_file_size":        10485760,
                 "max_per_event":        25,
                 "max_per_organization": 25,
                 "quality":              90,
             },
-            "documents": {
+            "video": {
+                "max_file_size":        10485760,
                 "max_per_event":        5,
                 "max_per_organization": 5,
             },
@@ -475,18 +489,18 @@ SUBSCRIPTION_PLANS = {
         },
         "accounts": {},
         "events": {
-            "upon_request_only":    False,
             "max_per_day":          1,
             "max_per_week":         None,
             "max_per_month":        None,
             "max_per_year":         None,
+            "upon_request_only":    False,
         },
         "organizations": {
-            "upon_request_only":    True,
             "max_per_day":          0,
             "max_per_week":         None,
             "max_per_month":        None,
             "max_per_year":         None,
+            "upon_request_only":    True,
         },
         "places": {},
     },
@@ -513,27 +527,16 @@ BOWER_INSTALLED_APPS = (
     "bootpag",
     "bootstrap#5.3.3",
     "bootstrap-maxlength",
-    # "bootstrap-rating",
     # "bootstrap-tagsinput",
-    # "bx-slider.js",
-    # "equalheight",
     "jquery#3.7.1",
-    # "jquery.inputmask",
     "jquery-colorbox",
     "jquery-file-upload#10.32.0",
     "jquery-popup-overlay#1.6.0",
-    # "jquery-shorten-js",
-    # "jquery-sticky",
     "jquery-ui#1.12.1",
-    # "jt.timepicker",
     "less.js#4.2.0",
-    # "modernizr",
     "moment#2.30.1",
     "noty#3.1.4",
     "readmore-js",
-    # "seiyria-bootstrap-slider",
-    # "smooth-scroll.js",
-    # "tablesorter",
     "underscore#1.13.6",
 )
 
@@ -1267,9 +1270,6 @@ UPLOADER_SETTINGS = {
             "tiff": "image/tiff",
             "webp": "image/webp",
         },
-        "MAX_FILE_SIZE":    10485760,
-        "MAX_FILE_NUMBER":  5,
-        "AUTO_UPLOAD":      True,
     },
     "documents": {
         "MIME_TYPES_MAP": {
@@ -1281,9 +1281,6 @@ UPLOADER_SETTINGS = {
             "rtf":  "application/rtf",
             "txt":  "text/plain",
         },
-        "MAX_FILE_SIZE":    10485760,
-        "MAX_FILE_NUMBER":  5,
-        "AUTO_UPLOAD":      True,
     },
     "images": {
         "MIME_TYPES_MAP": {
@@ -1296,9 +1293,6 @@ UPLOADER_SETTINGS = {
             "tiff": "image/tiff",
             "webp": "image/webp",
         },
-        "MAX_FILE_SIZE":    10485760,
-        "MAX_FILE_NUMBER":  5,
-        "AUTO_UPLOAD":      True,
     },
     "video": {
         "MIME_TYPES_MAP": {
@@ -1309,9 +1303,6 @@ UPLOADER_SETTINGS = {
             "ogv":  "video/ogg",
             "webm": "video/webm",
         },
-        "MAX_FILE_SIZE":    10485760,
-        "MAX_FILE_NUMBER":  5,
-        "AUTO_UPLOAD":      True,
     },
     "audio": {
         "MIME_TYPES_MAP": {
@@ -1323,32 +1314,36 @@ UPLOADER_SETTINGS = {
             "wav":  "audio/wav",
             "weba": "audio/webm",
         },
-        "MAX_FILE_SIZE":    10485760,
-        "MAX_FILE_NUMBER":  5,
-        "AUTO_UPLOAD":      True,
-    }
+    },
 }
-
+# -------------------------------------------------------------------------------------------------
 SUPPORTED_DEFAULTS = [key for key, val in UPLOADER_SETTINGS["default"]["MIME_TYPES_MAP"].items()]
 SUPPORTED_DEFAULTS_STR = ", ".join(SUPPORTED_DEFAULTS)
 SUPPORTED_DEFAULTS_STR_EXT =\
     ",".join([f".{key}" for key, val in UPLOADER_SETTINGS["default"]["MIME_TYPES_MAP"].items()])
 SUPPORTED_DEFAULTS_STR_REG =\
     "|".join([key for key, val in UPLOADER_SETTINGS["default"]["MIME_TYPES_MAP"].items()])
-
+# -------------------------------------------------------------------------------------------------
 SUPPORTED_DOCUMENTS = [key for key, val in UPLOADER_SETTINGS["documents"]["MIME_TYPES_MAP"].items()]
 SUPPORTED_DOCUMENTS_STR = ", ".join(SUPPORTED_DOCUMENTS)
 SUPPORTED_DOCUMENTS_STR_EXT =\
     ",".join([f".{key}" for key, val in UPLOADER_SETTINGS["documents"]["MIME_TYPES_MAP"].items()])
 SUPPORTED_DOCUMENTS_STR_REG =\
     "|".join([key for key, val in UPLOADER_SETTINGS["documents"]["MIME_TYPES_MAP"].items()])
-
+# -------------------------------------------------------------------------------------------------
 SUPPORTED_IMAGES = [key for key, val in UPLOADER_SETTINGS["images"]["MIME_TYPES_MAP"].items()]
 SUPPORTED_IMAGES_STR = ", ".join(SUPPORTED_IMAGES)
 SUPPORTED_IMAGES_STR_EXT =\
     ",".join([f".{key}" for key, val in UPLOADER_SETTINGS["images"]["MIME_TYPES_MAP"].items()])
 SUPPORTED_IMAGES_STR_REG =\
     "|".join([key for key, val in UPLOADER_SETTINGS["images"]["MIME_TYPES_MAP"].items()])
+# -------------------------------------------------------------------------------------------------
+SUPPORTED_VIDEO = [key for key, val in UPLOADER_SETTINGS["video"]["MIME_TYPES_MAP"].items()]
+SUPPORTED_VIDEO_STR = ", ".join(SUPPORTED_VIDEO)
+SUPPORTED_VIDEO_STR_EXT =\
+    ",".join([f".{key}" for key, val in UPLOADER_SETTINGS["video"]["MIME_TYPES_MAP"].items()])
+SUPPORTED_VIDEO_STR_REG =\
+    "|".join([key for key, val in UPLOADER_SETTINGS["video"]["MIME_TYPES_MAP"].items()])
 
 
 ###############################################################################